[cudapoa] added a couple of tests for verifying static and adaptive banded NW against full-alignment

+++ cudapoa/tests/Test_CudapoaGenerateMSA2.cu
@@ -117,9 +117,8 @@ TEST_F(MSATest, CudapoaMSA)
     auto spoa_msa = spoa_generate_multiple_sequence_alignments(sequences);
 
 #ifndef SPOA_ACCURATE
-    for (int32_t i = 0; i < spoa_msa.size(); i++)
+    for (size_t i = 0; i < spoa_msa.size(); i++)
     {
-
         std::string msa = cudapoa_msa[0][i];
         msa.erase(std::remove(msa.begin(), msa.end(), '-'), msa.end());
         ASSERT_EQ(msa, sequences[i]);
